Add timeouts and non blocking buzzer sequences
Timeout when in main "menu" to enter a low power state (thats a
todo) and when waiting in an inner menu to return to the main
menu.

The ability to run buzzer sequences with a non blocking call to
modify LEDs or game logic during a sound sequence was necesary.